requirements.txt
pydantic_ai
httpx
uvicorn
logfire
python-dotenv
